+++ scripts/travis/_includes.sh
@@ -1,14 +1,26 @@
 # shellcheck shell=bash
 
+#
+# Travis helpers
+#
 function fold_start() {
   export CURRENT_FOLD_NAME="$1"
-  travis_fold start "$CURRENT_FOLD_NAME"
-  travis_time_start
+
+  if [[ "${TRAVIS_COMMIT:-}" != "" ]]; then
+    travis_fold start "$CURRENT_FOLD_NAME"
+    travis_time_start "$CURRENT_FOLD_NAME"
+  else
+    echo "Starting $CURRENT_FOLD_NAME"
+  fi
 }
 
 function fold_end() {
-  travis_time_finish
-  travis_fold end "$CURRENT_FOLD_NAME"
+  if [[ "${TRAVIS_COMMIT:-}" != "" ]]; then
+    travis_time_finish "$CURRENT_FOLD_NAME"
+    travis_fold end "$CURRENT_FOLD_NAME"
+  else
+    echo "Done with $CURRENT_FOLD_NAME"
+  fi
 }
 
 # shellcheck disable=SC1090