+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - @iov/bns: Add `MultisignatureTx` type and `isMultisignatureTx` helper function
   to BNS.
+- @iov/bns: Support encoding/decoding multisignature transactions.
 - @iov/bns-governance: Add `TreasurySend` proposal type and export
   `TreasurySendOptions`.
 - @iov/bns-governance: Support `TreasurySend` proposals in